Run only if compiled successfully

+++ run.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-make
+if make
+then
 
 rm iso/isolinux/meow.c32
 cp meow.c32 iso/isolinux
 mkisofs -o output.iso -b isolinux/isolinux.bin -c isolinux/boot.cat -no-emul-boot -boot-load-size 4 -boot-info-table iso
 qemu-system-x86_64 -cdrom output.iso $@
+
+fi
